refactor(configuration): introduce /configuration component directory
