+++ pages/index.tsx
@@ -6,7 +6,7 @@ import { SitClass, SitSchedule } from "../types/sitTypes";
 import { ClassConfig, ClassPopularityIndex, NotificationsConfig } from "../types/rezervoTypes";
 import { arraysAreEqual } from "../utils/arrayUtils";
 import AppBar from "../components/AppBar";
-import MobileConfigUpdateBar from "../components/MobileConfigUpdateBar";
+import MobileConfigUpdateBar from "../components/configuration/MobileConfigUpdateBar";
 import { createClassPopularityIndex } from "../lib/popularity";
 import { DateTime } from "luxon";
 import { useUserConfig } from "../hooks/useUserConfig";
